+++ CHANGELOG.md
@@ -18,8 +18,10 @@ Types of changes:
 
 # Added
 - GitHub action to add a changelog reminder
+- List of documents
 - Test for validation of Swagger documentation
 - Added integration with Approval app. Thanks to @eneiluj
+- Endpoint to list LibreSign files
 
 # Changed
 - Increment of coverage on backend code
@@ -34,6 +36,7 @@ Types of changes:
 - Rendering files tab in Nextcloud 20 and 21
 - Invalid method name when validating if a file signature has already been requested
 - Tests autoload
+- Correction of loading class after clicking sign in application
 
 ## 2.3.0 - 2021-05-22